bpftool: Retrieve pinned maps/programs in read-only mode when relevant
In their talk at the Linux Plumbers Conference 2021 [0], Jakub and
Arthur exposed a number of user experience issues that they faced when
working with BPF. One, in particular, is rather easy to address in
bpftool: it's the fact that, when retrieving a file descriptor for a
pinned map, bpftool (through libbpf) attempts to open the pinned path
with read and write permissions. When the user can read but not write to
that path, the call fails, and the user is not able to list or dump
objects with bpftool.

Libbpf recently got a new function that allows to pass flags when
retrieving pinned objects. Let's use it to send the BPF_F_RDONLY flag
when relevant:

- Every time we retrieve a pinned program, given that we never modify
  them.
- Every time we retrieve a map with no intention to change it (this
  excludes updates or deletions of entries, for example).
- Never for BPF links, which need both read and write permissions.

[0] https://linuxplumbersconf.org/event/11/contributions/933/

Reported-by: Jakub Sitnicki <jakub@cloudflare.com>
Reported-by: Arthur Fabre <afabre@cloudflare.com>
Signed-off-by: Quentin Monnet <quentin@isovalent.com>