bpftool: Generate type lists in "bpftool cgroup/map help"
The help message displayed on "bpftool cgroup help" or "bpftool map
help" lists the attach types or map types, respectively, available for
the associated subcommands. This is useful to understand what types are
available. But the list was so far hardcoded in the source, and had to
be updated each time a new type is supported.

Via libbpf, we have access to the list of names for the types that we
want to print in these messages. So we can generate the list at runtime,
and save developers the trouble of adding new types to the list when
they add it to libbpf.

After this commit, the types displayed in the help messages are
the same as before, although they are listed in a different order.

Note that the man pages for bpftool still require the update. I don't
see a trivial way to update them automatically without either parsing
libbpf sources or making the build for docs rely on the bpftool binary,
none of which sounds ideal.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>